# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Xihe API
  version: 1.0.0
  description: Canonical HTTP contract for Xihe public and service APIs.
servers:
  - url: /
paths:
...
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Problem' } }
...
  /internal/v1/config/{layer}/{domain}:
    get:
      operationId: getInternalConfig
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: layer, required: true, schema: { type: string } }, { in: path, name: domain, required: true, schema: { type: string } }]
      responses: { '200': { description: Configuration map }, '401': { $ref: '#/components/responses/Problem' } }
...
  /api/v1/config/{domain}:
    get: { operationId: getConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: domain, required: true, schema: { type: string } }], responses: { '200': { description: Configuration map } } }
...
  /api/v1/models:
    get: { operationId: listModels, security: [{ bearerAuth: [] }], responses: { '200': { description: Models } } }
...
  /api/v1/sessions/{sessionId}/attachments:
    post: { operationId: uploadAttachments, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Upload result } } }
  /api/v1/sessions/{sessionId}/attachments/{fileId}:
    get: { operationId: getAttachment, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }, { in: path, name: fileId, required: true, schema: { type: string } }], responses: { '200': { description: Attachment metadata } } }
    delete: { operationId: deleteAttachment, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }, { in: path, name: fileId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
  /api/v1/files/{fileId}:
    get: { operationId: getFile, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: fileId, required: true, schema: { type: string } }], responses: { '200': { description: File content } } }
    delete: { operationId: deleteFile, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: fileId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /internal/v1/context/{sessionId}/events:
    get: { operationId: readContextEvents, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Events } } }
    post: { operationId: appendContextEvent, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '201': { description: Appended } } }
...
  /internal/v1/agent/models:
    get: { operationId: agentModels, security: [{ bearerAuth: [] }], responses: { '200': { description: Models } } }
...
  /internal/v1/context/{sessionId}/snapshot:
    get: { operationId: contextSnapshot, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Snapshot } } }
  /internal/v1/context/{sessionId}/events/batch:
    post: { operationId: appendContextEvents, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '201': { description: Appended } } }
  /internal/v1/context/{sessionId}/events/latest:
    get: { operationId: latestContextSequence, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Sequence } } }
  /internal/v1/context/{sessionId}/refresh-sources:
    post: { operationId: refreshContextSources, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '202': { description: Accepted } } }
  /internal/v1/context/{sessionId}/replay:
    post: { operationId: replayContext, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Replayed } } }
  /internal/v1/context/{sessionId}/compact:
    post: { operationId: compactContext, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Compacted } } }
  /internal/v1/context/{sourceSessionId}/fork:
    post: { operationId: forkContext, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sourceSessionId, required: true, schema: { type: string } }], responses: { '201': { description: Forked } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }